helm: explicit number of DB connections
Enriches `reana-db` helm template to define explicitly the max number
of allowed DB connections. The default value is 100 which may not be
enough for installations launching many concurrent
workflows. Increasing it to 300. Moreover, having the number
explicitly defined in the template will allow site administrators to
easily customise it further.