helm: set concurrency policy for all cronjobs
Sets a forbidding concurrency policy for all cronjobs so that they would
not be allowed to start if an old cronjob instance is still running.

Useful to prevent scheduling new cron job pods when the past cron job
pods are not yet finished e.g. due to troubles with shared filesystem.