feat(reana-dev): add create-branch option to git-upgrade command (#884)
This commit adds a --create-branch option to the git-upgrade command.
When used, it creates the branch locally from the upstream branch and
pushes it to the origin, if the branch exists in upstream. This option
is especially useful when upgrading maintenance branches that don't
exist locally or in your fork.

Co-authored-by: Alputer <alp.tuna.453@gmail.com>