feat(reana-dev): allow using Colima/K3s as an alternative to Kind (#894)
Add option `--kubernetes` to several `reana-dev` commands to allow using
Colima/K3s as an alternative to Kind. Useful on macOS where a VM is
necessary to run Docker anyway. Developing locally in Colima/K3s in the
`latest` cluster mode saves time, because the built container images are
natively present in the VM, and it is not necessary anymore to load and
reload Docker images to the Kind cluster.